# Implement multithread preprocess