Add flag: workers. Use multiprocess to parallelize Tokenizer.binarize.