@huihuifan 2018-06-13
Can you provide some numerical measurements of timing for binarization of wmt en-de on 1 worker, etc?