@paramecinm 2018-06-13
On my macbook with 1 core, 4 processors.

worker count | run time(s) | run time with out build dictionary | speed up | speed up (without build dictionary)
-- | -- | -- | -- | --
1 | 1734 | 1522 | 1 | 1
2 | 1235 | 1023 | 1.40405 | 1.48778
4 | 909 | 697 | 1.90759 | 2.18364
8 | 815 | 603 | 2.12761 | 2.52405

I am running experiments on dev server with more workers, and try to figure out if it is necessary to parallelize building dictionary.