@edunov 2018-06-14
As discussed offline, let's test performance on more cores say 20-40. And let's check if we need to parallelize dictionary construction.