@paramecinm 2018-06-14
I have parallelized dictionary build, and here is the new speed up on fb dev server, who have 24 cores.

worker_cnt | time | speedup
-- | -- | --
1 | 2263 | 1
12 | 591 | 3.8291
16 | 550 | 4.11455
24 | 517 | 4.37718
32 | 527 | 4.29412
48 | 532 | 4.25376