@bouvrie 2018-06-19
Thanks for posting this PR; it'd be a great addition! It looks like in the proposed approach all the workers  simultaneously attempt to scan the entire input file. You might also consider benchmarking an approach where the parent process (and only the parent) reads the file and deals groups of lines to the workers through shared memory.