@paramecinm 2018-06-19
@bouvrie Python's shared memory can only share fix length array with basic types. To share strings is not supported by default. Implementing producer-consumer model will also have overhead. Also I have recored the time for each part of the program, on WMT dataset, reading the whole file will only take very few seconds while binarizing and merging worker result will take hundreds of seconds. So I think it won't be worthy to optimize the file reading part.