@bouvrie 2018-06-20
@paramecinm Thanks, fair point on the file loading. I took a closer look at `Tokenizer.tokenize` and with the default `tokenize_line` it looks like most of the time is possibly spent in everything except the tokenizing (dict lookups, publishing to the `replaced` consumer). Perhaps this is the source of the contention/bottleneck in the many-thread setting?