@paramecinm 2018-06-20
@bouvrie Because the python shared memory in multiprocess doesn't support share complex structures, and using Python's data sharing api like Pipe to pass data is even slower than write data to file and read them in. So I choose to write down the partial result of each process and read them to merge in main process. The write and merge part takes a large amount of time. To further optimize the performance, I can only think of use the much powerful multithread in C. It will be great if you can find some better approach!