@edunov 2018-06-26
I wonder if we will have a better performance if master instead of waiting for workers to finish binarization will consume their output as they produce it. It can be done through piping. Right now, pickle.load() seems to kill a lot of performance. Also, it should save memory, as you won't need to keep everything in ram.